requirements.txt
alphashape==1.3.1
astroid==2.14.1
attrs==21.2.0
Automat==20.2.0
beautifulsoup4==4.10.0
blis==0.7.7
cachetools==4.2.4
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-log==0.4.0
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
cycler==0.11.0
cymem==2.0.6
Cython==0.29.30
descartes==1.1.0
dill==0.3.6
Distance==0.1.3
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
filelock==3.4.0
fonttools==4.28.5
ftfy==6.1.1
g2p-en==2.1.0
gdown==4.2.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-speech==2.11.1
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
h2==3.2.0
hpack==3.0.0
httplib2==0.20.2
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflect==5.6.0
isort==5.12.0
itemadapter==0.4.0
itemloaders==1.0.4
jarowinkler==1.0.5
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.3.2
langcodes==3.3.0
lazy-object-proxy==1.9.0
Levenshtein==0.18.1
libcst==0.3.21
-e git+https://gitlab.com/mikegringo/listenability-experiments.git@7c99424db09bf5a0218ea6011cb69cc487030211#egg=local_cainesap_syllabify&subdirectory=local_cainesap_syllabify
loguru==0.6.0
lxml==4.7.1
MarkupSafe==2.1.1
matplotlib==3.5.1
mccabe==0.7.0
msgpack==1.0.4
murmurhash==1.0.7
mypy-extensions==0.4.3
networkx==2.8.4
nltk==3.6.5
num2words==0.5.10
numpy==1.21.4
oauth2client==4.1.3
onnxruntime==1.7.0
packaging==21.3
pandas==1.3.5
parsel==1.6.0
pathy==0.6.1
patsy==0.5.2
Pillow==9.0.0
plac==1.3.5
platformdirs==3.0.0
praat-parselmouth==0.4.1
preshed==3.0.6
priority==1.3.0
Protego==0.1.16
proto-plus==1.19.8
protobuf==3.19.1
ptitprince==0.2.5
punctuator==0.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycparser==2.21
pydantic==1.8.2
PyDispatcher==2.0.5
pydub==0.25.1
pyenchant==3.2.2
PyHamcrest==2.0.3
pylint==2.16.1
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyphen==0.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
queuelib==1.6.2
rapidfuzz==2.1.0
regex==2021.11.10
requests==2.26.0
researchpy==0.3.2
rsa==4.7.2
Rtree==1.0.0
scikit-learn==1.0.2
scipy==1.7.3
Scrapy==2.5.1
seaborn==0.11.2
service-identity==21.1.0
Shapely==1.8.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
SpeechRecognition==3.8.1
srsly==2.4.3
statsmodels==0.13.2
termcolor==1.1.0
texterrors==0.4.2
textstat==0.7.2
tgt==1.4.4
Theano==1.0.5
thinc==8.0.16
threadpoolctl==3.0.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.62.3
trimesh==3.12.7
Twisted==21.7.0
typer==0.4.1
typing-inspect==0.7.1
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
w3lib==1.22.0
wasabi==0.9.1
wcwidth==0.2.6
wordfreq==3.0.3
wrapt==1.14.1
zope.interface==5.4.0
